feat: complete add model ticket and banner

+++ src/controllers/category.controller.ts
@@ -2,6 +2,7 @@ import { Response, Request } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interfaces";
 import CategoryModel, { categoryDAO } from "../models/category.model";
 import response from "../utils/response";
+import { isValidObjectId } from "mongoose";
 export default {
   async create(req: IReqUser, res: Response) {
     try {
@@ -56,7 +57,15 @@ export default {
   async findOne(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "category not found");
+      }
+
       const result = await CategoryModel.findById(id);
+      if (!result) {
+        return response.notFound(res, "category not found");
+      }
       response.success(res, result, "success find one category");
     } catch (error) {
       response.error(res, error, "failed find one category");
@@ -65,6 +74,11 @@ export default {
   async update(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "failed update category");
+      }
+
       const result = await CategoryModel.findByIdAndUpdate(id, req.body, {
         new: true,
       });
@@ -76,6 +90,11 @@ export default {
   async remove(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "failed update category");
+      }
+
       const result = await CategoryModel.findByIdAndDelete(id, {
         new: true,
       });
